tratando exceções de PUT e DELETE

+++ src/main/java/com/davidsolli/desafio3/services/ClientService.java
@@ -3,8 +3,11 @@ package com.davidsolli.desafio3.services;
 import com.davidsolli.desafio3.DTO.ClientDTO;
 import com.davidsolli.desafio3.entities.Client;
 import com.davidsolli.desafio3.repositories.ClientRepository;
+import com.davidsolli.desafio3.services.exceptions.DatabaseException;
 import com.davidsolli.desafio3.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -39,15 +42,27 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyToEntity(entity, dto);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyToEntity(entity, dto);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try{
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     public void copyToEntity(Client entity, ClientDTO dto) {
